Revert "add hobbies"
