+++ package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "dedent-js": "^1.0.1",
-    "gsap": "^3.12.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.11.0",
