+++ src/components/Button/Default.tsx
@@ -5,14 +5,8 @@ interface ButtonInterface {
 
 const DefaultButton = (props: ButtonInterface) => {
 	return (
-		<a
-			href={props.link}
-			className="bg-[#1C2B3A] py-[10px] px-[14px] rounded-md max-w-max block text-white hover:bg-[#607B96] transition-all duration-300"
-			target="_blank"
-			aria-label={props.text}
-		>
-			{props.text}
-		</a>
+		<a href={props.link} className="bg-[#1C2B3A] py-[10px] px-[14px] rounded-md max-w-max block text-white"
+		   target="_blank">{props.text}</a>
 	);
 };
