--- src/pages/About/Hobbies/Development/content.txt
@@ -1,39 +1,0 @@
-/**
-* @title About development
-*
-* @description
-* I am a passionate developer who likes to learn new things and improve my skills.
-* You can find my projects on projects page, or on my github.
-* I like to work in front-end and back-end, but i prefer back-end.
-*
-* @front
-* In front-end side, i use some technologies like :
-* - HTML
-* - CSS
-* - SASS
-* - Javascript
-* - ReactJS (TSX & JSX)
-* - TailwindCSS
-*
-* @back
-* In back-end side, i use some technologies like :
-* - PHP
-* - Symfony
-* - MySQL
-* - MongoDB
-* - ExpressJS
-* - Mongoose
-* - NodeJS
-* - NextJS
-*
-* @other
-* I would like to learn some tools like :
-* - Docker
-* - Kubernetes
-* - Github Actions
-* - CI/CD
-*
-* @future
-* For the future, i would like to be a fullstack developer, and learn some other languages of programmation
-* I would like to learn a .NET language too, like C#
-*/
