--- src/pages/About/Hobbies/Gaming/content.txt
@@ -1,23 +1,0 @@
-/**
-* @title About gaming
-*
-* @description
-* I am a passionate about gaming and i like to play to a lot of different type of games.
-*
-* @competitive
-* I'm competitive player, so i like play to :
-* - Counter-Strike
-* - Rocket League
-* - Escape From Tarkov
-* - Valorant
-*
-* @simulation
-* I also like to play simulation games like :
-* - Satisfactory
-* - Factorio
-* - Construction Simulator
-* - Foxhole
-*
-* @other
-* I like watching streams on Twitch if i don't play to games.
-*/
