+++ src/pages/About/Hobbies/hobbies.tsx
@@ -1,0 +1,9 @@
+const Hobbies = () => {
+	return (
+		<div>
+			<h1>Hobbies</h1>
+		</div>
+	)
+}
+
+export default Hobbies
